LEO-resilience Toolkit/LEO-resilience-toolkit/backend/app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
pydantic==2.11.7
pydantic-settings==2.10.1
pyyaml==6.0.2
skyfield==1.53
numpy==2.3.2
python-dateutil==2.9.0.post0
pytest==8.4.1
httpx==0.28.1

pandas==2.3.2
plotly==6.3.0
streamlit==1.49.0
pydeck==0.9.1
requests==2.32.5
